LayoutTests/platform/chromium/fast/forms/suggestion-picker/week-suggestion-picker-step-attribute-expected.txt

Tests that step attribute filters out suggestions as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
PASS entryValues().length is 6
PASS entryValues()[0] is "2012-01-21"
PASS entryValues()[1] is "2012-01-23"
PASS entryValues()[2] is "2012-01-25"
PASS entryValues()[3] is "2012-01-27"
PASS entryValues()[4] is "2012-01-29"
PASS entryValues()[5] is "@openCalendarPicker"
PASS entryValues().length is 3
PASS entryValues()[0] is "2012-01-21"
PASS entryValues()[1] is "2012-01-26"
PASS entryValues()[2] is "@openCalendarPicker"
PASS entryValues().length is 2
PASS entryValues()[0] is "2012-01-21"
PASS entryValues()[1] is "@openCalendarPicker"
PASS successfullyParsed is true

TEST COMPLETE
